move type-specific fns to types.go
