+++ sia/encoding.go
@@ -3,7 +3,6 @@ package sia
 import (
 	"encoding/binary"
 	"errors"
-	"math/big"
 	"reflect"
 )
 
@@ -197,60 +196,4 @@ func unmarshal(b []byte, val reflect.Value) (consumed int) {
 	}
 	panic("could not unmarshal type " + val.Type().String())
 	return
-}
-
-// MarshalSia implements the Marshaler interface for Signatures.
-func (s *Signature) MarshalSia() []byte {
-	if s.R == nil || s.S == nil {
-		return []byte{0, 0}
-	}
-	// pretend Signature is a tuple of []bytes
-	// this lets us use Marshal instead of doing manual length-prefixing
-	return Marshal(struct{ R, S []byte }{s.R.Bytes(), s.S.Bytes()})
-}
-
-// MarshalSia implements the Unmarshaler interface for Signatures.
-func (s *Signature) UnmarshalSia(b []byte) int {
-	// inverse of the struct trick used in Signature.MarshalSia
-	str := struct{ R, S []byte }{}
-	if Unmarshal(b, &str) != nil {
-		return 0
-	}
-	s.R = new(big.Int).SetBytes(str.R)
-	s.S = new(big.Int).SetBytes(str.S)
-	return len(str.R) + len(str.S) + 2
-}
-
-// MarshalSia implements the Marshaler interface for PublicKeys.
-func (pk *PublicKey) MarshalSia() []byte {
-	if pk.X == nil || pk.Y == nil {
-		return []byte{0, 0}
-	}
-	// see Signature.MarshalSia
-	return Marshal(struct{ X, Y []byte }{pk.X.Bytes(), pk.Y.Bytes()})
-}
-
-// MarshalSia implements the Unmarshaler interface for PublicKeys.
-func (pk *PublicKey) UnmarshalSia(b []byte) int {
-	// see Signature.UnmarshalSia
-	str := struct{ X, Y []byte }{}
-	if Unmarshal(b, &str) != nil {
-		return 0
-	}
-	pk.X = new(big.Int).SetBytes(str.X)
-	pk.Y = new(big.Int).SetBytes(str.Y)
-	return len(str.X) + len(str.Y) + 2
-}
-
-// MerkleRoot calculates the Merkle root hash of a SpendConditions object,
-// using the timelock, number of signatures, and the signatures themselves as leaves.
-func (sc *SpendConditions) MerkleRoot() Hash {
-	tlHash := HashBytes(Marshal(sc.TimeLock))
-	nsHash := HashBytes(Marshal(sc.NumSignatures))
-	pkHashes := make([]Hash, len(sc.PublicKeys))
-	for i := range sc.PublicKeys {
-		pkHashes[i] = HashBytes(Marshal(sc.PublicKeys[i]))
-	}
-	leaves := append([]Hash{tlHash, nsHash}, pkHashes...)
-	return MerkleRoot(leaves)
 }
